backend/sop_generator/requirements.txt
fastapi
python-dotenv
google-generativeai
uvicorn
pydantic
pydantic-settings
reportlab